decimal should be double

+++ CelestiaComponent/CelestiaDMS.h
@@ -27,7 +27,7 @@ namespace winrt::CelestiaComponent::implementation
         double HMSSeconds();
         double Decimal();
     private:
-        int32_t decimal;
+        double decimal;
     };
 }
